refactor(lotto): 유틸함수 제거
