--- src/Utils/Utils.js
@@ -1,31 +1,0 @@
-import readline from 'readline';
-
-export const MakeRandom = {
-  number(min, max) {
-    return Math.floor(Math.random() * (max - min + 1)) + min;
-  },
-};
-
-export const readLine = {
-  async(query) {
-    return new Promise((resolve, reject) => {
-      if (arguments.length !== 1) {
-        reject(new Error('arguments must be 1'));
-      }
-
-      if (typeof query !== 'string') {
-        reject(new Error('query must be string'));
-      }
-
-      const rl = readline.createInterface({
-        input: process.stdin,
-        output: process.stdout,
-      });
-
-      rl.question(query, input => {
-        rl.close();
-        resolve(input);
-      });
-    });
-  },
-};
